Rename Rollup plugin to "generate-json-folders"
